+++ utils/js/LogUtils.ts
@@ -6,9 +6,56 @@
  */
 import {Some} from '@xh/hoist/core';
 import {castArray, isString} from 'lodash';
+import store from 'store2';
 import {intersperse} from './LangUtils';
 
+/**
+ * Utility functions providing managed, structured logging to hoist apps.  Essentially a wrapper
+ * around the browser console supporting logging levels, timing, and miscellaneous hoist
+ * display conventions.
+ *
+ * Objects extending HoistBase need not import these functions directly, as they are available
+ * via delegates on `HoistBase`.
+ *
+ * Hoist sets its minimum severity level to 'info' by default.  This prevents performance or
+ * memory impacts that might result from verbose debug logging.  This can be adjusted by calling
+ * XH.logLevel from the console.
+ */
+
+/** Severity Level for log statement */
+export type LogLevel = 'error' | 'warn' | 'info' | 'debug';
+
+/** Object identifying the source of log statement.  Typically, a javascript class */
 export type LogSource = string | {displayName: string} | {constructor: {name: string}};
+
+/**
+ * Severity threshold for app.  Messages with less severity will be ignored. Default 'info'.
+ * @internal Applications should typically use `XH.logLevel` instead.
+ */
+export function getLogLevel() {
+    return _logLevel;
+}
+
+/**
+ * Adjust severity level of logging for lifetime of page or browser tab.
+ *
+ * Call this method from the console to adjust the log level for troubleshooting.
+ * @internal Applications should typically use `XH.setLogLevel()` instead.
+ */
+export function setLogLevel(level: LogLevel, persistInSessionStorage: boolean = false) {
+    level = level.toLowerCase() as LogLevel;
+    if (!['error', 'warn', 'info', 'debug'].includes(level)) {
+        console.error(`Invalid value for log level: '${level}'`);
+        return;
+    }
+    _logLevel = level;
+    if (persistInSessionStorage) {
+        store.session.set('xhLogLevel', level);
+    }
+    if (level != 'info') {
+        console.warn(`Client logging set to level '${level}'.`);
+    }
+}
 
 /**
  * Time and log execution of a function to `console.info()`.
@@ -72,19 +119,14 @@ export function logWarn(msgs: Some<unknown>, source?: LogSource) {
     return loggedDo(msgs, null, source, 'warn');
 }
 
-/** Parse a LogSource in to a canonical string label. */
-export function parseSource(source: LogSource): string {
-    if (!source) return null;
-    if (isString(source)) return source;
-    if (source['displayName']) return source['displayName'];
-    if (source.constructor) return source.constructor.name;
-    return null;
-}
-
 //----------------------------------
 // Implementation
 //----------------------------------
-function loggedDo<T>(messages: Some<unknown>, fn: () => T, source: LogSource, level: LogLevel) {
+function loggedDo<T>(messages: Some<unknown>, fn: () => T, source: LogSource, level: LogLevel): T {
+    if (_severity[level] < _severity[_logLevel]) {
+        return fn?.();
+    }
+
     let src = parseSource(source);
     let msgs = castArray(messages);
 
@@ -147,4 +189,20 @@ function writeLog(msgs: unknown[], src: string, level: LogLevel) {
     }
 }
 
-type LogLevel = 'error' | 'warn' | 'info' | 'debug';
+/** Parse a LogSource in to a canonical string label. */
+function parseSource(source: LogSource): string {
+    if (!source) return null;
+    if (isString(source)) return source;
+    if (source['displayName']) return source['displayName'];
+    if (source.constructor) return source.constructor.name;
+    return null;
+}
+
+//----------------------------------------------------------------
+// Initialization + Level/Severity support.
+// Initialize during parsing to make available immediately.
+//----------------------------------------------------------------
+let _logLevel: LogLevel = 'info';
+const _severity: Record<LogLevel, number> = {error: 3, warn: 2, info: 1, debug: 0};
+
+setLogLevel(store.session.get('xhLogLevel', 'info'));